Extract all of the transformer code into its own module
There's an import issue when `transformers` contains all of the
transformation logic.  I've moved it to its own module to avoid.  This
change keeps all of the individual transformers in their own module and
all of the transformation code in a separate module.

This is definitely a candidate for renaming to avoid the singular-plural
confusion that could happen.