Quick and dirty way to generate an ID for each person
This doesn't allow tracking people across name changes for sure, and if
you include departments and job titles, we lose the ability to track
them across department moves and title changes.

The idea is to start asking departments for some sort of identifier that
is unique to a person and will stay with them as they move around.
Hopefully we're able to get that, but that won't allow moves into
completely different areas (El Paso County to City of El Paso or UTEP,
for example).